% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peptide-mhc.R
\name{smmPMBEC}
\alias{smmPMBEC}
\title{Peptide-MHC Binding Prediction}
\usage{
smmPMBEC(peptides = c("SLYNTVATL", "SYFPEITHI"), mhc = "HLA-A02:01",
  output.IC50 = TRUE)
}
\arguments{
\item{peptides}{vector of strings containing the peptides for which to predict
binding.}

\item{mhc}{string or vector of strings identifying the MHC molecules. See
\code{supportedMHCs} for allowed values.}

\item{output.IC50}{whether to output the IC50 value itself (default) or its
base10-logarithm.}
}
\description{
Predicts peptide-MHC binding using the stabilized matrix method (SMM)
algorithm with a specifically constructed amino acid substitution matrix.
See \code{citation(package='PeptideMHC')} for the reference.
}

