% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peptide-mhc.R
\name{supportedMHCs}
\alias{supportedMHCs}
\title{Supported MHC Molecules and Peptide Lengths.}
\usage{
supportedMHCs()
}
\value{
A named list. The names are the MHC molecules and the list entries
are integer vectors containing the supported peptide lengths for each MHC.
}
\description{
Supported MHC Molecules and Peptide Lengths.
}
\examples{
## Which MHC molecules are supported?
names( supportedMHCs() )
## Which peptide lengths are supported for HLA-A02:01?
supportedMHCs()[["HLA-A02:01"]]
}

